requirements.txt
click==8.1.7
Pillow==10.1.0
typer==0.9.0
typing_extensions==4.8.0
